Create a new file to contain report generation logic
